+++ src/Jobs/TimeoutWorkflowStep.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace JustSteveKing\WorkflowEngine\Jobs;
 
 use Illuminate\Bus\Queueable;
+use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
@@ -13,6 +14,7 @@ use JustSteveKing\WorkflowEngine\Exceptions\WorkflowNotFoundException;
 
 final class TimeoutWorkflowStep implements ShouldQueue
 {
+    use Dispatchable;
     use InteractsWithQueue;
     use Queueable;
     use SerializesModels;